+++ client/src/components/project/FilesManager.vue
@@ -23,6 +23,7 @@
         <q-tooltip>Create a new volume from selected data</q-tooltip>
       </q-btn>
     </q-toolbar>
+
     <div class="col-xs-12 col-sm-12 col-md-12">
       <FilesViewer :data="data"/>
 
@@ -34,76 +35,9 @@
         transition-show="slide-up"
         transition-hide="slide-down"
       >
-        <q-card class="bg-white text-dark">
-          <q-bar>
-            <div class="text-h5">Create New Volume</div>
-            <q-space />
-            <q-btn dense flat icon="close" v-close-popup>
-              <q-tooltip>Close</q-tooltip>
-            </q-btn>
-          </q-bar>
-
-          <q-card-section>
-            <div class="row">
-              <div class="col-sm-5 col-md-5 col-lg-5 col-xl-5">
-                <h6 class="no-margin">Shared Data</h6>
-                <q-input
-                  ref="filter"
-                  dense
-                  v-model="filter"
-                  label="Filter"
-                  class="q-px-md"
-                >
-                  <template v-slot:append>
-                    <q-icon
-                      v-if="filter !== ''"
-                      name="clear"
-                      class="cursor-pointer"
-                      @click="resetFilter"
-                    />
-                  </template>
-                </q-input>
-                <q-tree
-                  :nodes="sharedData"
-                  node-key="label"
-                  tick-strategy="leaf"
-                  :ticked.sync="ticked"
-                  :filter="filter"
-                />
-              </div>
-              <div
-                class="col-sm-2 col-md-2 col-lg-2 col-xl-2 self-center text-center q-px-sm"
-              >
-                <q-icon size="24px" name="arrow_forward" />
-                <div>Select files on the left to create a new volume</div>
-                <q-icon size="24px" name="arrow_forward" />
-              </div>
-              <div class="col-sm-5 col-md-5 col-lg-5 col-xl-5">
-                <h6 class="no-margin">New Volume</h6>
-                <q-input v-model="newVolumeName" dense />
-                <q-tree
-                  :nodes="newVolumeChildren"
-                  node-key="label"
-                  default-expand-all
-                />
-              </div>
-            </div>
-            <div class="row justify-end">
-              <q-btn
-                type="submit"
-                @click="createVolume"
-                label="Create Volume"
-                class="q-mt-md"
-                color="teal"
-              >
-                <template v-slot:loading>
-                  <q-spinner-facebook />
-                </template>
-              </q-btn>
-            </div>
-          </q-card-section>
-        </q-card>
+        <AddNewVolume :data.sync="data" :volumesDialog.sync="volumesDialog" />
       </q-dialog>
+
       <!-- File Uploaded Dialog  -->
       <q-dialog v-model="fileUploadDialog" position="standard">
         <FileUploader />
@@ -117,36 +51,23 @@ import { uid } from 'quasar'
 import { gql } from '@apollo/client'
 
 import FileUploader from '../upload/FileUploader'
+import AddNewVolume from './modals/AddNewVolume'
 import FilesViewer from './FilesViewer'
 
 export default {
   name: 'FilesManager',
   components: {
     FileUploader,
-    FilesViewer
+    FilesViewer,
+    AddNewVolume
   },
   data: () => ({
-    ticked: [],
-    filter: '',
-    sharedData: [],
     data: [], // Tree
-    newVolumeChildren: [],
     volumesDialog: false,
     fileUploadDialog: false,
-    maximizedToggle: true,
-    newVolumeName: 'New Volume'
+    maximizedToggle: true
   }),
   watch: {
-    // whenever question changes, this function will run
-    ticked () {
-      this.newVolumeChildren = []
-      this.ticked.forEach((checkedEle) => {
-        this.newVolumeChildren.push({
-          label: checkedEle,
-          icon: 'insert_drive_file'
-        })
-      })
-    },
     '$route': 'fetchData'
   },
   async created () {
@@ -187,31 +108,7 @@ export default {
           children: [...files]
         }
       )
-    },
-    resetFilter () {
-      this.filter = ''
-      this.$refs.filter.focus()
-    },
-    createVolume () {
-      const newVolume = {
-        label: this.newVolumeName,
-        icon: 'folder',
-        children: this.newVolumeChildren
-      }
-      this.data.push(newVolume)
-      this.ticked = []
-      this.volumesDialog = false
     }
   }
 }
 </script>
-
-<style>
-.profile-border {
-  border: 5px solid white!important;
-}
-.dataView {
-  max-width: 700px;
-  margin: auto;
-}
-</style>
